Merge develop → main: Authentication, employee management, and tasks

+++ backend/src/main/java/cz/zcu/kiv/caretracker/dto/performedTask/PerformedTaskDTO.java
@@ -17,7 +17,7 @@ import java.util.List;
 public class PerformedTaskDTO {
     private Long id;
     private LocalDateTime date;
-    private Integer unitCount;
+    private Double unitCount;
     private String notes;
 
     private ClientSummaryDTO client;
